Release wakelock for incoming SMS.
- In the current implementation, a wakelock is acquired when SMS is
received in IdleState. The wakelock is released 3 seconds after
returning back to IdleState. As the statemachine will be back in
IdleState right after SMS processing is complete, there is no need to
hold the wakelock for another 3 seconds.

- Add a change to hold wakelock for 3 seconds as a worst case timeout
rather than for normal processing of SMS. Also maintain a wakelock
counter to make sure that wakelock is released when statemachine is
idle.

CRs-Fixed: 875160
Change-Id: I826603d7e5d8ea5ced3d331ddb17618be0390ffe