Notify remaining attempts for SIM Lock upon bootup.
The current logic is in such a way that when there
is Incorrect password entered from UI Modem throws
exception with number of possible re-attempts.
In order to support getting re-attempts upon bootup
(with out entering any password) introduced dummy
pin which is passed to modem and get the success
result along with the number of re-attempts.

Since there is no exception from modem in this case,
The condition is not met. Hence number of re-attempts
is not notified to UI.

Exception condition has been removed so that proper re-
maining attempts will be passed to SIM lock screen.

Change-Id: I5d2ab37cef028781723bf587650ca9156f59d3b9
CRs-Fixed: 912687